Handle update error
